Hack Listview keyboard navigation for better UX
I would have preferred to find some "stock" mechanism to do
this but in the end I couldn't fin a combination that works.

The effect is using the arrow keys will navigate the selected item
in the list without requiring an additional arrow press to switch focus
from the search to the list. The tab and shift+tab keys are used as
alias to up/down instead of the wonky behaviour of tab nav builtin
to ListView.